Tyuiu.KultyshevaEA.Sprint2.Task5.V14.Lib/DataService.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;
using tyuiu.cources.programming.interfaces.Sprint2;

namespace Tyuiu.KultyshevaEA.Sprint2.Task5.V14.Lib
{
    public class DataService : ISprint2Task5V14
    {
        public string FindDayName(int k, int d)
        {
            string res;
            int firstDayOfWeek = 1;

            int remainder = (k - 1) % 7;

            int dayofweek = (firstDayOfWeek + remainder) % 7;
            switch (dayofweek)
            {
                case 1:
                    res = "Понедельник";
                    break;
                case 2:
                    res = "Вторник";
                    break;
                case 3:
                    res = "Среда";
                    break;
                case 4:
                    res = "Четверг";
                    break;
                case 5:
                    res = "Пятница";
                    break;
                case 6:
                    res = "Суббота";
                    break;
                case 7:
                    res = "Воскресенье";
                    break;
                default:
                    throw new ArgumentException($"День должен быть от 1 до 365. Значение {k}" );

            }
            return res;

        }
    }
}
